geospatial_app/requirements.txt
databricks-sdk>=0.46.0

# Core Streamlit framework
streamlit>=1.28.0
streamlit-folium>=0.15.0
folium>=0.14.0

# Data manipulation
pandas>=2.0.0

# Optional: For enhanced UI
plotly>=5.17.0

# KeplerGL
keplergl
streamlit-keplergl
